feat(Categoria, Tamanhos): fechando essas routes para somento os desevolvedores cadrastados na empresa de id 1 possam opera

+++ src/categorias/categorias.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateCategoriaDto } from './dto/create-categoria.dto';
 import { Categoria } from './entities/categoria.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -11,8 +11,11 @@ export class CategoriasService {
     private categoriaRepository: Repository<Categoria>,
   ) { }
 
-  async create(createCategoriaDto: CreateCategoriaDto): Promise<Categoria> {
-    return await this.categoriaRepository.save(createCategoriaDto);
+  async create(createCategoriaDto: CreateCategoriaDto, dateUser: { id: number, name: string, empresa: string }): Promise<Categoria> {
+    if (dateUser.empresa && dateUser.empresa === process.env.MASTER_NAME) {
+      return await this.categoriaRepository.save(createCategoriaDto);
+    }
+    throw new BadRequestException("Voce nao tem autorizacao para inserir uma categoria nova!!!")
   }
 
   async findAll(): Promise<Categoria[]> {
@@ -23,15 +26,18 @@ export class CategoriasService {
     );
   }
 
-  async remove(id: number): Promise<void> {
-    const categoria = await this.categoriaRepository.findOne({
-      where: { id: id }
-    });
-
-    if (!categoria) {
-      throw new NotFoundException("Essa categoria nao existe!!");
+  async remove(id: number, dateUser: { id: number, name: string, empresa: string }): Promise<void> {
+    if (!dateUser.empresa && dateUser.empresa !== process.env.MASTER_NAME) {
+      const categoria = await this.categoriaRepository.findOne({
+        where: { id: id }
+      });
+  
+      if (!categoria) {
+        throw new NotFoundException("Essa categoria nao existe!!");
+      }
+  
+      await this.categoriaRepository.remove(categoria);
     }
-
-    await this.categoriaRepository.remove(categoria);
+    throw new BadRequestException("Voce nao tem autorizacao para remove essa categoria!!!")
   }
 }
